Day 6

+++ 2015/solutions/c/src/advent/solution.c
@@ -53,7 +53,7 @@ char *read_file(char *path) {
     fread(buffer, sizeof(char), size_buffer, stream);
     buffer[size_buffer] = '\0';
     /* trim trailing whitespace */
-    for (char *p = buffer + size_buffer - 1; isspace(*p); p--)
+    for (char *p = buffer + size_buffer - 1; isspace((int) *p); p--)
         *p = '\0';
     fclose(stream);
     return buffer;
